+++ aospy/utils/vertcoord.py
@@ -4,7 +4,7 @@ import logging
 import numpy as np
 import xarray as xr
 
-from ..constants import grav
+from .._constants import GRAV_EARTH
 from .. import internal_names
 
 
@@ -160,7 +160,7 @@ def vert_coord_name(arr):
 def int_dp_g(arr, dp):
     """Mass weighted integral."""
     return integrate(arr, to_pascal(dp, is_dp=True),
-                     vert_coord_name(dp)) / grav.value
+                     vert_coord_name(dp)) / GRAV_EARTH
 
 
 def dp_from_p(p, ps, p_top=0., p_bot=1.1e5):