+++ docs/whats-new.rst
@@ -11,6 +11,10 @@ v0.3 (unreleased)
 Breaking Changes
 ~~~~~~~~~~~~~~~~
 
+- Deprecate ``Constant`` class and ``constants.py`` module.
+  Physical constants used internally by aospy are now stored
+  in ``_constants.py`` (fixes :issue:`50` via :pull:`223`).
+  By `Micah Kim <https://github.com/micahkim23>`_.
 - Deprecate ``Units`` class, so now the ``units`` attribute of the
   ``Var`` class is a string. (fixes :issue:`50` via :pull:`222`).
   By `Micah Kim <https://github.com/micahkim23>`_.